fix: исправить замечания
по замечанию на 49 строке card.js: слушатель в данной строке добавляется
